fix merge conflict, apply fixes from upstream

+++ crates/anvil/src/eth/backend/mem/mod.rs
@@ -3151,10 +3151,10 @@ impl TransactionValidator for Backend {
             }
 
             // Ensure the tx does not exceed the max blobs per block.
-            if blob_count > MAX_BLOBS_PER_BLOCK_DENCUN as usize {
+            if blob_count > MAX_BLOBS_PER_BLOCK_DENCUN {
                 return Err(InvalidTransactionError::TooManyBlobs(
                     blob_count,
-                    MAX_BLOBS_PER_BLOCK_DENCUN as usize,
+                    MAX_BLOBS_PER_BLOCK_DENCUN,
                 ))
             }
